[refactor] 잘못된 패키지명 변경(FileStorage -> file_storage)
