+++ src/main/java/flab/project/common/file_storage/UploadedFileUrls.java
@@ -1,4 +1,4 @@
-package flab.project.common.fileStorage;
+package flab.project.common.file_storage;
 
 import java.util.Set;
 import java.util.stream.Collectors;
