+++ src/main/java/flab/project/domain/user/controller/UserController.java
@@ -2,7 +2,7 @@ package flab.project.domain.user.controller;
 
 import static flab.project.utils.AccessManagementUtil.assertUserIdOwner;
 
-import flab.project.common.fileStorage.UploadedFileUrl;
+import flab.project.common.file_storage.UploadedFileUrl;
 import flab.project.common.annotation.LoggedInUserId;
 import flab.project.config.baseresponse.SuccessResponse;
 import flab.project.domain.user.model.Profile;
