+++ src/main/java/flab/project/domain/user/facade/UserFacade.java
@@ -1,7 +1,7 @@
 package flab.project.domain.user.facade;
 
-import flab.project.common.fileStorage.FileStorage;
-import flab.project.common.fileStorage.UploadedFileUrl;
+import flab.project.common.file_storage.FileUploader;
+import flab.project.common.file_storage.UploadedFileUrl;
 import flab.project.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,11 +12,11 @@ import static flab.project.domain.file.enums.FileType.PROFILE_IMAGE;
 @Service
 public class UserFacade {
 
-    private final FileStorage fileStorage;
+    private final FileUploader fileUploader;
     private final UserService userService;
 
     public UploadedFileUrl updateProfileImage(long userId) {
-        UploadedFileUrl uploadedFileUrl = fileStorage.generatePreSignedUrl(userId, PROFILE_IMAGE);
+        UploadedFileUrl uploadedFileUrl = fileUploader.generatePreSignedUrl(userId, PROFILE_IMAGE);
 
         userService.updateProfileImage(userId, uploadedFileUrl.getContentUrl());
